[add] ユーザと投稿記事との関連付けと権限制御

+++ app/Http/Controllers/ArticleController.php
@@ -47,6 +47,7 @@ class ArticleController extends Controller
             'body' => 'required'
         ]);
         $article = new Article();
+        $article->user_id = \Auth::id();
         $article->title = $request->title;
         $article->body = $request->body;
         $article->save();
@@ -76,6 +77,7 @@ class ArticleController extends Controller
     public function edit(Article $article)
     {
         //
+        $this->authorize($article);
         $data = ['article' => $article];
         return view('articles.edit', $data);
     }
@@ -90,6 +92,7 @@ class ArticleController extends Controller
     public function update(Request $request, Article $article)
     {
         //
+        $this->authorize($article);
         $this->validate($request, [
             'title' => 'required|max:255',
             'body' => 'required'
@@ -109,6 +112,7 @@ class ArticleController extends Controller
     public function destroy(Article $article)
     {
         //
+        $this->authorize($article);
         $article->delete();
         return redirect(route('articles.index'));
     }
